R08/T02/feature/register_page2 | Mobx
- Agora o valor inicial das categorias da homePage não é mais Açaí. Este problema foi resolvido definindo novos @observable e @action para categorias em `register_store.dart`
- Também movi @observable e @action referente ao plano para `register_card.dart`
