+++ lib/app/modules/register/restaurant/register_store.dart
@@ -2,14 +2,34 @@ import 'package:mobx/mobx.dart';
 
 part 'register_store.g.dart';
 
-class RestaurantRegisterStore = _RestaurantRegisterStoreBase with _$RestaurantRegisterStore;
-abstract class _RestaurantRegisterStoreBase with Store {
+class RestaurantRegisterStore = _RestaurantRegisterStoreBase
+    with _$RestaurantRegisterStore;
 
+abstract class _RestaurantRegisterStoreBase with Store {
   @observable
   int value = 0;
 
   @action
   void increment() {
     value++;
-  } 
+  }
+
+  @observable
+  var selectedCategory = 'Açaí';
+
+  @action
+  void setSelectedCategory(String? selection) {
+    selectedCategory = selection ?? 'Açaí';
+  }
+
+  @observable
+  var availablePlans = ['Plano Básico', 'Plano Entrega'];
+
+  @observable
+  var selectedPlan = 'Plano Básico';
+
+  @action
+  void setSelectedPlan(String? selection) {
+    selectedPlan = selection ?? availablePlans.first;
+  }
 }
